make page-not-found more lazy loading
